+++ src/Resources/contao/languages/de/tl_settings.php
@@ -1,0 +1,24 @@
+<?php if (!defined('TL_ROOT')) die('You cannot access this file directly!');
+
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (C) 2005-2013 Leo Feyer
+ *
+ * @package   fen
+ * @author    Frank Hoppe
+ * @license   GNU/LGPL
+ * @copyright Frank Hoppe 2013
+ */
+
+/**
+ * Name
+ */
+$GLOBALS['TL_LANG']['tl_settings']['insert_verein_legend']              = 'Insert-Tag verein';
+
+/**
+ * fields
+ */
+$GLOBALS['TL_LANG']['tl_settings']['insert_verein_replaces'] = array('Ersetzungen', 'Ersetzungen für Vereinsnamen definieren');
+$GLOBALS['TL_LANG']['tl_settings']['insert_verein_search'] = array('Suchen nach', 'Leerzeichen am Anfang oder Ende bitte als + angeben');
+$GLOBALS['TL_LANG']['tl_settings']['insert_verein_replace'] = array('Ersetzen durch', 'Leerzeichen am Anfang oder Ende bitte als + angeben');